refactor(app): modularize performance harness
