+++ packages/app/package.json
@@ -18,13 +18,14 @@
     "build": "vite build",
     "serve": "vite preview",
     "test": "bun run test:unit && bun run test:virtualizer",
-    "test:unit": "bun test --only-failures --preload ./happydom.ts ./src ./e2e/performance/chrome-trace-write.test.ts ./e2e/performance/timeline/session-tab-switch-metrics.test.ts",
+    "test:unit": "bun test --only-failures --preload ./happydom.ts ./src",
     "test:virtualizer": "bun test --conditions=browser --preload ./happydom.ts ./test-browser/solid-virtual.test.ts",
     "test:unit:watch": "bun test --watch --preload ./happydom.ts ./src",
     "test:e2e": "playwright test",
     "test:e2e:local": "playwright test",
     "test:e2e:ui": "playwright test --ui",
     "test:e2e:report": "playwright show-report e2e/playwright-report",
+    "test:bench": "bun test ./e2e/performance/chrome-trace-write.test.ts ./e2e/performance/timeline/session-tab-switch-metrics.test.ts && OPENCODE_PERFORMANCE=1 playwright test e2e/performance",
     "trace:stats": "devtools-tracing stats",
     "trace:inp": "devtools-tracing inp",
     "trace:selectors": "devtools-tracing selector-stats"
